// src/app/components/configuracion/interfaces/Usuario.ts
export interface Usuario {
  id: number;
  nombre: string;
  apellido: string;
  dui: string;
  nit: string;
  nrc: string;
  correo: string;
  telefono: string;
  pais: string;
  departamento: string;
  municipio: string;
  complemento: string;
  credencialCvcp: string;
  periodoInicio: string;
  periodoFin: string;
}